requirements.txt
Flask
Werkzeug
nltk
numpy
Pillow
scikit-learn
